Create first python file
This is the first file created in the cs50x course.

I have run into a problem where the code as written will not run if I use the terminal. It will run on Replit or using the "run" button in the top right of VScode. But if I try to run it with the terminal I get this output:

    MacBook-Pro-2:repos Shurtugal$ cd cs50x/
    MacBook-Pro-2:cs50x Shurtugal$ python hello.py
    What's your name? kieran
    Traceback (most recent call last):
    File "hello.py", line 2, in <module>
    name = input("What's your name? ")
    File "<string>", line 1, in <module>
    NameError: name 'kieran' is not defined

I don't have an explanation for this yet, and will just resort to using the "run" button for testing.